New ArtifactStore to support keeping files from prev. extracted artifact
The ExtractedArtifactStore allows to define either a specific directory containing the extracted artifact
files for a distribution or to download and extract an artifact into a fixed path directory.

The ArtifactStoreBuilder now allows to configure if the extracted artifact files should be kept. It is made
sure to not conflict with useCache.